requirements.txt
# NexusOS Blockchain Hub - Dependencies

# Core Framework
streamlit>=1.28.0
streamlit-autorefresh>=1.0.1

# Web & API
flask>=3.0.0
flask-cors>=4.0.0
python-socketio>=5.10.0
requests>=2.31.0

# Blockchain & Cryptography
web3>=6.11.3
eth-account>=0.10.0
eth-keys>=0.4.0
hexbytes>=0.3.1
cryptography>=41.0.7
bcrypt>=4.1.2
passlib>=1.7.4

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.4
pandas>=2.1.4
numba>=0.58.1
scikit-learn>=1.3.2
scikit-optimize>=0.9.0

# Visualization
plotly>=5.18.0

# Graph & Network
networkx>=3.2.1

# PDF Processing
pypdf2>=3.0.1

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.3
scikit-learn
scipy
flask-socketio
werkzeug
psycopg2-binary
socketio
